+++ CHANGELOG.md
@@ -4,10 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Support the label `restarter.stackable.tech/ignore` on ConfigMaps and Secrets and the annotations
+  `restarter.stackable.tech/ignore-configmap.x` and `restarter.stackable.tech/ignore-secret.x` on
+  StatefulSets to exclude ConfigMaps and Secrets from the restarter controller ([#410]).
+
 ### Changed
 
 - Document Helm deployed RBAC permissions and remove unnecessary permissions ([#412]).
+- Bump stackable-operator to version 0.110.0 ([#410]).
 
+[#410]: https://github.com/stackabletech/commons-operator/pull/410
 [#412]: https://github.com/stackabletech/commons-operator/pull/412
 
 ## [26.3.0] - 2026-03-16